update to kld clamping
Former-commit-id: 667a9260b7993f2a38de6bea8e1e62d9c723a3b2

+++ vampire/modules/vae/logistic_normal.py
@@ -18,7 +18,7 @@ class LogisticNormal(VAE):
                  mean_projection: FeedForward,
                  log_variance_projection: FeedForward,
                  decoder: FeedForward,
-                 kld_clamp: Optional[int] = None,
+                 kld_clamp: Optional[float] = None,
                  apply_batchnorm: bool = False,
                  z_dropout: float = 0.2) -> None:
         super(LogisticNormal, self).__init__(vocab)
@@ -86,7 +86,9 @@ class LogisticNormal(VAE):
         mu, sigma = params["mean"], params["variance"]  # pylint: disable=C0103
         negative_kl_divergence = 1 + torch.log(sigma ** 2) - mu ** 2 - sigma ** 2
         if self._kld_clamp:
-            negative_kl_divergence = torch.clamp(negative_kl_divergence, min=-self._kld_clamp, max=self._kld_clamp)
+            negative_kl_divergence = torch.clamp(negative_kl_divergence,
+                                                 min=-self._kld_clamp,
+                                                 max=self._kld_clamp)
         negative_kl_divergence = 0.5 * negative_kl_divergence.sum(dim=-1)  # Shape: (batch, )
         return negative_kl_divergence
